#81 | Fix for performance issues in form editing
Fixes lags in expanding/collapsing, changing group/element data.
This commit fixes it by using React.memo in components so they do not
have to rerender on each change that's not of relevence to them. Also
had to replace Object.assign with immer as Object.assign does shallow
copy and that screws up areEqual function of memoized components. Immer
allows to work with immutable state in very easy way.