#251 | Fix sync performance
Fixing the main query run during the sync by filtering by id instead of
uuid. For example, we were getting the program uuid from client and on
server filtering program enrolments by program uuid
and that was causing the query to be very slow. Now we will still be
getting program uuid from client but on the server we will get an id for
that program and filter by the id. Filtering by the id is way faster.